+++ src/app/core/models/custom-inputs-option.model.ts
@@ -1,0 +1,11 @@
+export interface CustomInputsOption {
+  id?: number;
+  mandatory?: boolean;
+  name?: string;
+  options?: string[];
+  placeholder?: string;
+  prefix?: string;
+  type?: string;
+  label: string;
+  value: string;
+}
